Convert SMPS music and SFX to SMPS2ASM.
We know from the SMPS source code that the original codebase used ASM
music and SFX as well. This eliminates a bunch of obscure binary
blobs from the disassembly.

Instead of using the original ASM format from SMPS's source code,
I've opted for Flamewing's SMPS2ASM.

One reason for this is that the format reveals bugs more clearly. An
example of this is the coordination flags being implemented as macros
instead of constants (as they are in the original format), allowing
parameters to be explicitly denoted and have sanity checks performed
on them.

Additionally, SMPS2ASM is designed with portability in mind, using
its macros to convert things such as coordination flags and FM voices
to suit multiple SMPS driver revisions. Sonic hackers like to port
songs between games, so this should constitute a usability
improvement.

Having these ASM-ified SMPS files in this disassembly is also a nice
way of ensuring their accuracy to the original bytecode, due to the
disassembly's focus on producing an exact copy of the original ROM.

Notably, SMPS2ASM allows a number of bugs in the original music data
to be seen. For instance, did you know that the credits music has an
extra little jingle after its end that plays on the PSG2 channel? A
buggy coordination flag renders it inaudible. SMPS2ASM allows you to
see this buggy command for yourself.

The upstream repository for SMPS2ASM and its converted music and SFX
are at this URL:
https://github.com/flamewing/flamedriver